Move User & VoteHistoryItem to structs
