+++ src/pages/submissions.rs
@@ -1,7 +1,7 @@
 use super::base::{get_base_template, BaseTemplate};
-use crate::auth::User;
 use crate::db::get_submissions;
 use crate::error::Error;
+use crate::structs::User;
 use crate::util::human_relative_time;
 
 use askama::Template;
